+++ cogs/owner.py
@@ -1,9 +1,22 @@
+#  Copyright 2019 Allan Galarza
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 import inspect
 import os
 import platform
 import textwrap
 import traceback
-from collections import defaultdict
 from contextlib import redirect_stdout
 from distutils.version import StrictVersion
 
@@ -66,6 +79,53 @@ class Owner(commands.Cog, CogUtils):
         for admin in guild_admins:
             await admin.send("{0}\n\t-{1.mention}".format(content, ctx.author))
         await ctx.send("Message sent to "+join_list(["@"+a.name for a in guild_admins], ", ", " and "))
+
+    @checks.owner_only()
+    @commands.command()
+    async def editmessage(self, ctx: NabCtx, message_link: str, *, new_content: str):
+        """Edits a bot message with new content based on its JSON representation.
+
+        JSON based on the [Discord API](https://discordapp.com/developers/docs/resources/channel#embed-object) format.
+        A visualizer can be seen [here](https://leovoel.github.io/embed-visualizer/).
+        """
+        try:
+            data = json.loads(self.cleanup_code(new_content))
+        except json.JSONDecodeError:
+            return await ctx.error("New content is not a valid json string.")
+
+        if not isinstance(data, dict):
+            return await ctx.error("New content is not a valid json string.")
+
+        content = data.get("content")
+        embed = None
+        if "embed" in data:
+            if "timestamp" in data["embed"] and isinstance(data["embed"]["timestamp"], str):
+                data["embed"]["timestamp"] = data["embed"]["timestamp"].replace('Z', '')
+            embed = discord.Embed.from_dict(data["embed"])
+
+        guild_id, channel_id, message_id = parse_message_link(message_link)
+        if channel_id is None or message_id is None:
+            return await ctx.error("That's not a valid message link.")
+        if guild_id is None:
+            return await ctx.error("I can't edit private messages.")
+        guild: discord.Guild = ctx.bot.get_guild(guild_id)
+        if guild is None:
+            return await ctx.error("I'm not in the guild the message belongs to.")
+        channel: discord.TextChannel = guild.get_channel(channel_id)
+        if channel_id is None:
+            return await ctx.error("I can't find the channel the message belongs to.")
+        try:
+            message: discord.Message = await channel.fetch_message(message_id)
+        except discord.HTTPException:
+            return await ctx.error("I couldn't find the message.")
+        if message.author != ctx.me:
+            return await ctx.error("I can only edit my own messages.")
+
+        try:
+            await message.edit(content=content, embed=embed)
+            await ctx.message.add_reaction("✅")
+        except discord.HTTPException:
+            return await ctx.error("I could edit the message. Maybe the content is malformed or exceeds limits.")
 
     @checks.owner_only()
     @commands.command()
@@ -297,7 +357,6 @@ class Owner(commands.Cog, CogUtils):
         except discord.HTTPException:
             return await ctx.error("Error uploading file. It is currently not possible to read the current log file.")
 
-
     @commands.command(usage="<old world> <new world>")
     @checks.owner_only()
     async def merge(self, ctx: NabCtx, old_world: str, new_world: str):
@@ -441,12 +500,81 @@ class Owner(commands.Cog, CogUtils):
 
     @checks.owner_only()
     @commands.command()
+    async def messagejson(self, ctx: NabCtx, message_link: str):
+        """Shows the json representation of a message.
+
+        JSON based on the [Discord API](https://discordapp.com/developers/docs/resources/channel#embed-object) format.
+        A visualizer can be seen [here](https://leovoel.github.io/embed-visualizer/).
+        """
+        guild_id, channel_id, message_id = parse_message_link(message_link)
+        if channel_id is None or message_id is None:
+            return await ctx.error("That's not a valid message link.")
+        if guild_id is None:
+            return await ctx.error("I can't check private messages.")
+        guild: discord.Guild = ctx.bot.get_guild(guild_id)
+        if guild is None:
+            return await ctx.error("I'm not in the guild the message belongs to.")
+        channel: discord.TextChannel = guild.get_channel(channel_id)
+        if channel_id is None:
+            return await ctx.error("I can't find the channel the message belongs to.")
+        try:
+            message: discord.Message = await channel.fetch_message(message_id)
+        except discord.HTTPException:
+            return await ctx.error("I couldn't find the message.")
+
+        data = dict()
+        if message.content:
+            data["content"] = message.content
+        if message.embeds:
+            # We only care about the first, regular embed.
+            embed: discord.Embed = message.embeds[0]
+            if embed.type == "rich":
+                data["embed"] = embed.to_dict()
+                del data["embed"]["type"]
+
+        await ctx.send(f"```json\n{json.dumps(data, indent=1)}```")
+
+    @checks.owner_only()
+    @commands.command()
     async def ping(self, ctx: NabCtx):
         """Shows the bot's response times."""
         resp = await ctx.send('Pong! Loading...')
         diff = resp.created_at - ctx.message.created_at
         await resp.edit(content=f'Pong! That took {1000*diff.total_seconds():.1f}ms.\n'
                                 f'Socket latency is {1000*self.bot.latency:.1f}ms')
+
+    @checks.owner_only()
+    @commands.command()
+    async def sendmessage(self, ctx: NabCtx, channel: Optional[discord.TextChannel] = None, *, json_content: str):
+        """Sends a message based on its JSON representation.
+
+        JSON based on the [Discord API](https://discordapp.com/developers/docs/resources/channel#embed-object) format.
+        A visualizer can be seen [here](https://leovoel.github.io/embed-visualizer/)."""
+        if channel is None:
+            channel = ctx.channel
+
+        try:
+            data = json.loads(self.cleanup_code(json_content))
+        except json.JSONDecodeError:
+            return await ctx.error("Content is not a valid json string.")
+
+        if not isinstance(data, dict):
+            return await ctx.error("Content is not a valid json string.")
+
+        content = data.get("content")
+        embed = None
+        if "embed" in data:
+            if "timestamp" in data["embed"] and isinstance(data["embed"]["timestamp"], str):
+                data["embed"]["timestamp"] = data["embed"]["timestamp"].replace('Z', '')
+            embed = discord.Embed.from_dict(data["embed"])
+
+        try:
+            await channel.send(content, embed=embed)
+            await ctx.message.add_reaction("✅")
+        except discord.Forbidden:
+            await ctx.error("I don't have the right permissions to send a message there.")
+        except discord.HTTPException:
+            await ctx.error("I couldn't send your message, the content might be malformed or exceed limits.")
 
     @checks.owner_only()
     @commands.command()