+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 - 🐛 Fixed bug
 - ❌ Removed feature
 
+## Version 2.4.0 (Unreleased)
+- ✔ New owner command `/sendmessage` to send a message based on its JSON representation.
+- ✔ New owner command `/editmessage` to edit a bot's message's content based on its json representation.
+- ✔ New owner command `/jsonmessage` to see the JSON representation of any message.
+- 🔧 `/quote` now accepts message links for faster quoting.
+- 🔧 Updated TibiaWiki database to reflect new rent prices and article updates.
+- 🔧 Reworked `/bosses` command.
+- 🔧 News tickers can now be disabled using `/settings newstickers`.
+- 🔧 Memory optimizations, prefixes are now read from memory instead of the database.
+- 🐛 Fixed RAM percentage inconsistencies in `/about`.
+
 ## Version 2.3.0 (2019-04-19)
 - ✔ New subcommand `/unregistered guild`, checks which members of a guild are not registered in the server.
 - ✔ New owner command `/logs` to upload log files.